feito: reorganização do diretório
